step4 - 파일 분리
